Refactor Timetable imports and directory structure
Moved Timetable components from 'Timetable.jsx' subdirectory to 'Timetable' for improved organization. Updated all relevant imports in MyAPPanel components to reflect the new paths. Also fixed a conditional in Timetable.jsx to properly check for multiple languages.
